requirements.txt
google-adk
mcp
requests
python-dotenv
streamlit
